+++ packages/utilities/src/index.ts
@@ -3,3 +3,4 @@ export * from "./object-utils";
 export * from "./context-utils";
 export * from "./assertion";
 export * from "./error-utils";
+export * from "./wave-utils";