+++ packages/utilities/src/wave-utils.ts
@@ -1,0 +1,3 @@
+import { WaveConfig } from "@vastly/types";
+
+export const defineConfig = (options: WaveConfig) => options;