+++ services/api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # api-service
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @vastly/utils@0.3.1
+
 ## 0.1.0
 
 ### Minor Changes